Improve offer matching algorithm
Match on the tab title as well as the tab url.
This introduced complications because the tab title is not always loaded until a lot later.
To get around this, the core matching logic is called again if the tab is in an "undefined" state but there are no offers found yet.